move extension type definition to extension.h so that the Convert method can be used elsewhere
