+++ r/src/extension-impl.cpp
@@ -25,69 +25,7 @@
 #include <arrow/extension_type.h>
 #include <arrow/type.h>
 
-// A wrapper around arrow::ExtensionType that allows R to register extension
-// types whose Deserialize, ExtensionEquals, and Serialize methods are
-// in meaningfully handled at the R level. At the C++ level, the type is
-// already serialized to minimize calls to R from C++.
-//
-// Using a std::shared_ptr<> to wrap a cpp11::sexp type is unusual, but we
-// need it here to avoid calling the copy constructor from another thread,
-// since this might call into the R API. If we don't do this, we get crashes
-// when reading a multi-file Dataset.
-class RExtensionType : public arrow::ExtensionType {
- public:
-  RExtensionType(const std::shared_ptr<arrow::DataType> storage_type,
-                 std::string extension_name, std::string extension_metadata,
-                 std::shared_ptr<cpp11::environment> r6_class,
-                 std::thread::id creation_thread)
-      : arrow::ExtensionType(storage_type),
-        extension_name_(extension_name),
-        extension_metadata_(extension_metadata),
-        r6_class_(r6_class),
-        creation_thread_(creation_thread) {}
-
-  std::string extension_name() const { return extension_name_; }
-
-  bool ExtensionEquals(const arrow::ExtensionType& other) const;
-
-  std::shared_ptr<arrow::Array> MakeArray(std::shared_ptr<arrow::ArrayData> data) const;
-
-  arrow::Result<std::shared_ptr<arrow::DataType>> Deserialize(
-      std::shared_ptr<arrow::DataType> storage_type,
-      const std::string& serialized_data) const;
-
-  std::string Serialize() const { return extension_metadata_; }
-
-  std::string ToString() const;
-
-  std::unique_ptr<RExtensionType> Clone() const;
-
-  cpp11::environment r6_class() const { return *r6_class_; }
-
-  cpp11::environment r6_instance(std::shared_ptr<arrow::DataType> storage_type,
-                                 const std::string& serialized_data) const;
-
-  cpp11::environment r6_instance() const {
-    return r6_instance(storage_type(), Serialize());
-  }
-
- private:
-  std::string extension_name_;
-  std::string extension_metadata_;
-  std::string cached_to_string_;
-  std::shared_ptr<cpp11::environment> r6_class_;
-  std::thread::id creation_thread_;
-
-  arrow::Status assert_r_thread() const {
-    if (std::this_thread::get_id() == creation_thread_) {
-      return arrow::Status::OK();
-    } else {
-      return arrow::Status::ExecutionError("RExtensionType <", extension_name_,
-                                           "> attempted to call into R ",
-                                           "from a non-R thread");
-    }
-  }
-};
+#include "extension.h"
 
 bool RExtensionType::ExtensionEquals(const arrow::ExtensionType& other) const {
   // Avoid materializing the R6 instance if at all possible, since this is slow
@@ -158,6 +96,20 @@ std::string RExtensionType::ToString() const {
   cpp11::function instance_ToString(instance[".ToString"]);
   cpp11::sexp result = instance_ToString();
   return cpp11::as_cpp<std::string>(result);
+}
+
+cpp11::sexp RExtensionType::Convert(const std::shared_ptr<arrow::Array>& array) {
+  cpp11::environment instance = r6_instance();
+  cpp11::function instance_Convert(instance[".array_as_vector"]);
+  cpp11::sexp array_sexp = cpp11::to_r6<arrow::Array>(array, "ExtensionArray");
+  return instance_Convert(array_sexp);
+}
+
+cpp11::sexp RExtensionType::Convert(const std::shared_ptr<arrow::ChunkedArray>& array) {
+  cpp11::environment instance = r6_instance();
+  cpp11::function instance_Convert(instance[".chunked_array_as_vector"]);
+  cpp11::sexp array_sexp = cpp11::to_r6<arrow::ChunkedArray>(array, "ChunkedArray");
+  return instance_Convert(array_sexp);
 }
 
 std::unique_ptr<RExtensionType> RExtensionType::Clone() const {
